[9386] Store some config values in Rate emun.

+++ src/game/AuctionHouseMgr.cpp
@@ -292,7 +292,7 @@ void AuctionHouseMgr::LoadAuctionItems()
         return;
     }
 
-    barGoLink bar( result->GetRowCount() );
+    barGoLink bar( (int)result->GetRowCount() );
 
     uint32 count = 0;
