+++ src/game/GMTicketMgr.cpp
@@ -48,7 +48,7 @@ void GMTicketMgr::LoadGMTickets()
         return;
     }
 
-    barGoLink bar( result->GetRowCount() );
+    barGoLink bar( (int)result->GetRowCount() );
 
     uint32 count = 0;
