+++ src/game/GameEventMgr.cpp
@@ -121,7 +121,7 @@ void GameEventMgr::LoadFromDB()
     uint32 count = 0;
 
     {
-        barGoLink bar( result->GetRowCount() );
+        barGoLink bar( (int)result->GetRowCount() );
         do
         {
             ++count;
@@ -187,7 +187,7 @@ void GameEventMgr::LoadFromDB()
     else
     {
 
-        barGoLink bar( result->GetRowCount() );
+        barGoLink bar( (int)result->GetRowCount() );
         do
         {
             Field *fields = result->Fetch();
@@ -233,7 +233,7 @@ void GameEventMgr::LoadFromDB()
     else
     {
 
-        barGoLink bar( result->GetRowCount() );
+        barGoLink bar( (int)result->GetRowCount() );
         do
         {
             Field *fields = result->Fetch();
@@ -281,7 +281,7 @@ void GameEventMgr::LoadFromDB()
     else
     {
 
-        barGoLink bar( result->GetRowCount() );
+        barGoLink bar( (int)result->GetRowCount() );
         do
         {
             Field *fields = result->Fetch();
@@ -338,7 +338,7 @@ void GameEventMgr::LoadFromDB()
     else
     {
 
-        barGoLink bar( result->GetRowCount() );
+        barGoLink bar( (int)result->GetRowCount() );
         do
         {
             Field *fields = result->Fetch();
@@ -382,7 +382,7 @@ void GameEventMgr::LoadFromDB()
     else
     {
 
-        barGoLink bar2( result->GetRowCount() );
+        barGoLink bar2( (int)result->GetRowCount() );
         do
         {
             Field *fields = result->Fetch();
