+++ src/game/PoolManager.cpp
@@ -418,7 +418,7 @@ void PoolManager::LoadFromDB()
 
     uint32 count = 0;
 
-    barGoLink bar(result->GetRowCount());
+    barGoLink bar((int)result->GetRowCount());
     do
     {
         ++count;
@@ -456,7 +456,7 @@ void PoolManager::LoadFromDB()
     else
     {
 
-        barGoLink bar2(result->GetRowCount());
+        barGoLink bar2((int)result->GetRowCount());
         do
         {
             Field *fields = result->Fetch();
@@ -518,7 +518,7 @@ void PoolManager::LoadFromDB()
     else
     {
 
-        barGoLink bar2(result->GetRowCount());
+        barGoLink bar2((int)result->GetRowCount());
         do
         {
             Field *fields = result->Fetch();
@@ -587,7 +587,7 @@ void PoolManager::LoadFromDB()
     else
     {
 
-        barGoLink bar2( result->GetRowCount() );
+        barGoLink bar2( (int)result->GetRowCount() );
         do
         {
             Field *fields = result->Fetch();
